add test for parameter list index, reorganize member and program indices tests
