+++ runtime/ast/parameter.go
@@ -18,10 +18,6 @@
 
 package ast
 
-import (
-	"sync"
-)
-
 type Parameter struct {
 	Label          string
 	Identifier     Identifier
@@ -39,39 +35,4 @@ func (p Parameter) EffectiveArgumentLabel() string {
 		return p.Label
 	}
 	return p.Identifier.Identifier
-}
-
-type ParameterList struct {
-	once                    sync.Once
-	Parameters              []*Parameter
-	_parametersByIdentifier map[string]*Parameter
-	Range
-}
-
-// EffectiveArgumentLabels returns the effective argument labels that
-// the arguments of a call must use:
-// If no argument label is declared for parameter,
-// the parameter name is used as the argument label
-//
-func (l *ParameterList) EffectiveArgumentLabels() []string {
-	argumentLabels := make([]string, len(l.Parameters))
-
-	for i, parameter := range l.Parameters {
-		argumentLabels[i] = parameter.EffectiveArgumentLabel()
-	}
-
-	return argumentLabels
-}
-
-func (l *ParameterList) ParametersByIdentifier() map[string]*Parameter {
-	l.once.Do(l.initialize)
-	return l._parametersByIdentifier
-}
-
-func (l *ParameterList) initialize() {
-	parametersByIdentifier := make(map[string]*Parameter, len(l.Parameters))
-	for _, parameter := range l.Parameters {
-		parametersByIdentifier[parameter.Identifier.Identifier] = parameter
-	}
-	l._parametersByIdentifier = parametersByIdentifier
 }
